ISSUE.

+++ DEVICE/INCLUDE/GRAPHICS/BINDABLES/GFX_vertexbuffer.h
@@ -22,12 +22,12 @@ public:
 
 	int GetIndexCount() const;
 
-	void UpdateVertexBuffer(ID3D11DeviceContext*, float, DirectX::XMFLOAT3 a, DirectX::XMFLOAT3 b, DirectX::XMFLOAT3 c);
+	void UpdateVertexBuffer(ID3D11DeviceContext*, int, DirectX::XMFLOAT3 a, DirectX::XMFLOAT3 b, DirectX::XMFLOAT3 c);
 
 	int GetVertexCount() const;
 
 private:
-	void Sierpinsky(float, DirectX::XMFLOAT3, DirectX::XMFLOAT3, DirectX::XMFLOAT3, ID3D11DeviceContext*);
+	void Sierpinsky(int, DirectX::XMFLOAT3, DirectX::XMFLOAT3, DirectX::XMFLOAT3, ID3D11DeviceContext*);
 
 private:
 	Microsoft::WRL::ComPtr<ID3D11Buffer> m_vertexBuffer;
